requirements.txt
asgiref==3.7.2
attrs==23.2.0
cffi==1.16.0
channels==4.0.0
coverage==7.5.3
cryptography==42.0.5
distlib==0.3.8
Django==5.0.2
django-clearcache==1.2.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
filelock==3.14.0
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
method_decorator==0.1.3
platformdirs==4.2.1
psycopg2-binary
pycparser==2.21
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
setuptools==70.0.0
sqlparse==0.4.4
uritemplate==4.1.1
virtualenv==20.26.2
watchdog==4.0.1